Add doc text to exposed apis
Add a response object for wrapping CURL responses in a more standardised manner.

Rename `generateSignature` to `signData` as it signs the data and does more than simply generates the signature based off data.

Expose and rename `sign` to `generateSignature`. This can be reused for validating webhooks sent from Checkout. etc.